refactor: 필드- 누락된 출력문 추가 / 테스트 코드 패키지로 분리 / 클래스 명 변경 / No Args - default 생성자 제거
