+++ src/main/java/lotto/domain/LottoRandom.java
@@ -5,13 +5,8 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 public class LottoRandom {
-    private final Random random;
     private static final int LOTTO_NUMBER_COUNT = 6;
     private static final int LOTTO_MAX_NUMBER = 45;
-
-    public LottoRandom(){
-        this.random = new Random();
-    }
 
     public ArrayList<Integer> createRandomNumbers(){
         List<Integer> numList = IntStream.range(1, LOTTO_MAX_NUMBER + 1)
