+++ src/main/java/lotto/domain/LottoTicket.java
@@ -3,7 +3,7 @@ package lotto.domain;
 import java.util.ArrayList;
 
 public class LottoTicket {
-    private static final int LOTTO_TICKET_SIZE = 6;
+    public static final int LOTTO_TICKET_SIZE = 6;
     public static final int LOTTO_LOWER_BOUND = 1;
     public static final int LOTTO_UPPER_BOUND = 45;
     private final ArrayList<Integer> lottoNumbers;
